restructure container path
