+++ web/src/components/atoms/Columns.js
@@ -1,8 +1,5 @@
 import styled from 'styled-components'
-import Chip from '@material-ui/core/Chip'
-import React from 'react'
 import Box from '@material-ui/core/Box'
-import { bps } from 'ui/responsive'
 
 export default styled(Box)`
   && {
