+++ web/src/components/atoms/Divider.js
@@ -1,0 +1,11 @@
+import { Divider } from '@material-ui/core'
+import styled from 'styled-components'
+
+export default styled(Divider)`
+  && {
+    background-color: #ececec;
+    width: 100%;
+    margin-top: 30px;
+    margin-bottom: 20px;
+  }
+`
