+++ web/src/components/district/DCCAOverview.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
 import styled from 'styled-components'
-import Box from '@material-ui/core/Box'
 import Typography from '@material-ui/core/Typography'
 import Paper from '@material-ui/core/Paper'
 import IconButton from '@material-ui/core/IconButton'
@@ -8,7 +7,7 @@ import ArrowBackIcon from '@material-ui/icons/ArrowBack'
 import ArrowForwardIcon from '@material-ui/icons/ArrowForward'
 import { Tag } from '../atoms/Tag'
 import { Button } from '@material-ui/core'
-import DistrictNewVoterChartContainer from 'components/DistrictNewVoterChartContainer'
+import DistrictNewVoterChartContainer from 'components/containers/DistrictNewVoterChartContainer'
 import Columns from 'components/atoms/Columns'
 import Rows from 'components/atoms/Rows'
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
